payment integer bug fixed

+++ EMUICTProject/ViewController/ConfirmOrderViewController.swift
@@ -15,7 +15,7 @@ class ConfirmOrderViewController: UIViewController,UINavigationControllerDelegat
     var bid:String!
     var sid:String!
     var product = [Product]()
-    var totalPrice = 0.00
+    var totalPrice = 0
     var totalMoney:String!
     
     @IBOutlet weak var tableView: UITableView!
@@ -54,7 +54,7 @@ class ConfirmOrderViewController: UIViewController,UINavigationControllerDelegat
 
         query.observe(.value) { (snapshot) in
             self.product.removeAll()
-            self.totalPrice = 0.00
+            self.totalPrice = 0
             for child in snapshot.children.allObjects as! [DataSnapshot] {
 
                 if let value = child.value as? NSDictionary {
@@ -63,7 +63,7 @@ class ConfirmOrderViewController: UIViewController,UINavigationControllerDelegat
                     let quant = value["Quantity"] as? String
                     let prices = value["Price"] as? String
 
-                    self.totalPrice = self.totalPrice + Double(prices!)!
+                    self.totalPrice = self.totalPrice + Int(prices!)!
                     print(self.totalPrice)
                     self.price.text = String(self.totalPrice)
